#18 Add dummy scaffold for components, add empty dashboard view, fix location of accidentially moved App.js and App.test.js
