+++ client/src/App.js
@@ -3,7 +3,7 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import PropTypes from 'prop-types';
-import logo from '../assets/images/everyMealLogo.png';
+import logo from './assets/images/everyMealLogo.png';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
@@ -12,7 +12,7 @@ import IconButton from '@material-ui/core/IconButton';
 import MenuIcon from '@material-ui/icons/Menu';
 import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
-import SignUp from './Signup.js';
+import SignUp from './views/Signup.js';
 
 
 /* Defining the basic layout of the application */
