+++ client/src/components/Footer.js
@@ -1,0 +1,4 @@
+"use strict";
+
+import React from 'react';
+
