+++ client/src/components/Header.js
@@ -1,0 +1,3 @@
+"use strict";
+
+import React from 'react';
