+++ client/src/components/Page.js
@@ -1,0 +1,4 @@
+"use strict";
+
+import React from 'react';
+
