+++ client/src/index.js
@@ -5,10 +5,11 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import { Provider } from 'react-redux';
 import configureStore from './Store/configureStore';
-import App from './views/App';
+import App from './App';
 import * as serviceWorker from './serviceWorker';
 import {Route, Link, BrowserRouter as Router} from "react-router-dom";
-import Signup from './views/Signup'
+import Signup from './views/Signup';
+import Dashboard from './views/Dashboard';
 
 const store = configureStore();
 
@@ -18,6 +19,7 @@ const routing = (
       <div>
         <Route exact path="/" component= {App} />
         <Route path="/signup" component= {Signup} />
+        <Route path="/dashboard" component= {Dashboard} />
       </div>
     </Router>
   </Provider>
