+++ client/src/views/Dashboard.js
@@ -1,0 +1,14 @@
+"use strict";
+
+import React from 'react';
+
+import Header from '../components/Header';
+import Footer from '../components/Footer';
+
+export class Dashboard extends React.Component {
+    render() {
+        return 0;
+    }
+}
+
+export default Dashboard;
